Implemented timeout policy in ConnectionManager. Fixed issue with acceptance of messages not sent to the device. Fixed issue related to duplication of messages.

+++ include/config/lora_protocol.h
@@ -1,10 +1,11 @@
 #pragma once
 #include <chrono>
+#include <cstdint>
 
 namespace CertSense {
 constexpr size_t message_buffer_capacity = 10;
 constexpr size_t max_window_size = 3;
-constexpr size_t first_seq = 0;
+constexpr uint16_t first_seq = 0;
 constexpr double timeout_scale = 1.5;
 constexpr uint8_t ack = 0;
 constexpr uint8_t arq = 1;
